Module stronghold_communication::actor::firewall

Structs

FirewallConfiguration

PermissionSum
The sum of allowed permissions. This is using the same concepts as e.g. permission values in Unix systems.

PermissionValue
The permission value for request variants. It is a bit that is set at a certain index, therefore the value is always a power of 2.

Enums

FirewallPermission
Permission that is set in the Firewall. In case of FirewallPermission::Restricted, only selected variants in a enum are allowed, the VariantPermission of the request message is used for each individual request to validate it.

FirewallRule
Configure the firewall.

RequestDirection
The direction of a [CommunicationRequest::RequestMsg] that firewall receives.

Traits

ToPermissionVariants
Convert an element to implement permissions.

VariantPermission
The permission value for the different variants of an enum. This allows permitting specific variants of an enum while prohibiting others. In structs or unions, it should default to PermissionValue(1)