Struct stronghold_communication::actor::CommunicationActor

pub struct CommunicationActor<Req, Res, T, U> where
    Req: MessageEvent,
    Res: MessageEvent,
    T: Message + From<Req>,
    U: Message + From<FirewallRequest<Req>>,  {
    swarm_tx: Option<UnboundedSender<(CommunicationRequest<Req, T>, Sender)>>,
    swarm_task_config: Option<(Keypair, CommunicationConfig<Req, T, U>, BehaviourConfig)>,
    poll_swarm_handle: Option<RemoteHandle<()>>,
    marker: PhantomData<Res>,
}

Actor for the communication to a remote peer over the swarm.

Sends the CommunicationRequests it receives over the swarm to the corresponding remote peer and forwards incoming request to the client provided in the CommunicationConfig. Before forwarding any requests, a FirewallRequest is send to the firewall and the Request will only be forwarded if a FirewallResponse::Accept was returned.

If remote peers should be able to dial the local system, a CommunicationRequest::StartListening has to be sent to the CommunicationActor.

use libp2p::identity::Keypair;
use riker::actors::*;
use serde::{Deserialize, Serialize};
use stronghold_communication::{
    actor::{firewall::OpenFirewall, CommunicationActor, CommunicationConfig},
    behaviour::BehaviourConfig,
};

#[derive(Debug, Clone, Serialize, Deserialize)]
pub enum Request {
    Ping,
}

#[derive(Debug, Clone, Serialize, Deserialize)]
pub enum Response {
    Pong,
}

// blank client actor without any logic
#[derive(Clone, Debug)]
struct ClientActor;

impl ActorFactory for ClientActor {
    fn create() -> Self {
        ClientActor
    }
}

impl Actor for ClientActor {
    type Msg = Request;

    fn recv(&mut self, _ctx: &Context<Self::Msg>, _msg: Self::Msg, sender: Sender) {}
}

let local_keys = Keypair::generate_ed25519();
let sys = ActorSystem::new().expect("Init actor system failed.");
let keys = Keypair::generate_ed25519();
let firewall = sys
    .actor_of::<OpenFirewall<Request>>("firewall")
    .expect("Init firewall actor failed.");
let client = sys
    .actor_of::<ClientActor>("client")
    .expect("Init client actor failed.");
let actor_config = CommunicationConfig::new(client, firewall);
let behaviour_config = BehaviourConfig::default();
let comms_actor = sys
    .actor_of_args::<CommunicationActor<_, Response, _, _>, _>(
        "communication",
        (local_keys, actor_config, behaviour_config),
    )
    .expect("Init communication actor failed.");

Fields
swarm_tx: Option<UnboundedSender<(CommunicationRequest<Req, T>, Sender)>>swarm_task_config: Option<(Keypair, CommunicationConfig<Req, T, U>, BehaviourConfig)>poll_swarm_handle: Option<RemoteHandle<()>>marker: PhantomData<Res>
Implementations

impl<Req, Res, T, U> CommunicationActor<Req, Res, T, U> where
Req: MessageEvent,
Res: MessageEvent,
T: Message + From<Req>,
U: Message + From<FirewallRequest<Req>>,

fn send_swarm_task(
&mut self,
msg: CommunicationRequest<Req, T>,
sender: Sender
) -> Result<(), SendError>