Struct stronghold_communication::behaviour::P2PNetworkBehaviour

pub struct P2PNetworkBehaviour<T: MessageEvent, U: MessageEvent> {
    mdns: Mdns,
    identify: Identify,
    msg_proto: RequestResponse<MessageCodec<T, U>>,
    peers: BTreeMap<PeerId, Multiaddr>,
    events: Vec<P2PEvent<T, U>>,
    response_channels: BTreeMap<String, ResponseChannel<U>>,
}

The P2PNetworkBehaviour determines the behaviour of the p2p-network. It combines the following protocols from libp2p

mDNS for peer discovery within the local network

identify-protocol to receive identifying information of the remote peer

RequestResponse Protocol for sending generic request T and response U messages

The P2PNetworkBehaviour itself is only useful when a new Swarm is created for it, and this swarm is the entry point for all communication to remote peers, and contains the current state.

The P2PNetworkBehaviour implements a custom poll method that creates P2PEvents from the events of the different protocols, it can be polled with the next() or next_event() methods of libp2p::swarm::ExpandedSwarm.

Fields
mdns: Mdnsidentify: Identifymsg_proto: RequestResponse<MessageCodec<T, U>>peers: BTreeMap<PeerId, Multiaddr>events: Vec<P2PEvent<T, U>>response_channels: BTreeMap<String, ResponseChannel<U>>
Implementations

impl<T: MessageEvent, U: MessageEvent> P2PNetworkBehaviour<T, U>

pub fn init_swarm(
local_keys: Keypair
) -> Result<Swarm<P2PNetworkBehaviour<T, U>>, BehaviourError>

Creates a new P2PNetworkBehaviour and returns the swarm for it. The returned Swarm<P2PNetworkBehaviour> is the entry point for all communication with remote peers, i.g. to send requests and responses. Additionally to the methods of the P2PNetworkBehaviour there is a range of libp2p::Swarm functions for swarm interaction, like dialing a new peer, that can be used.

Example

use libp2p::{
    core::{Multiaddr, PeerId},
    identity::Keypair,
    request_response::{RequestId, RequestResponseEvent, ResponseChannel},
};
use serde::{Deserialize, Serialize};
use stronghold_communication::behaviour::{MessageEvent, P2PNetworkBehaviour};

#[derive(Debug, Clone, Serialize, Deserialize)]
pub enum Request {
    Ping,
}

#[derive(Debug, Clone, Serialize, Deserialize)]
pub enum Response {
    Pong,
}

let local_keys = Keypair::generate_ed25519();
let mut swarm = P2PNetworkBehaviour::<Request, Response>::init_swarm(local_keys).unwrap();

fn poll<TEv>(
&mut self,
_cx: &mut Context<'_>,
_params: &mut impl PollParameters
) -> Poll<NetworkBehaviourAction<TEv, P2PEvent<T, U>>>

pub fn add_peer(&mut self, peer_id: PeerId, addr: Multiaddr)

pub fn remove_peer(&mut self, peer_id: &PeerId) -> Option<Multiaddr>

pub fn get_peer_addr(&self, peer_id: &PeerId) -> Option<&Multiaddr>

pub fn get_all_peers(&self) -> Vec<(&PeerId, &Multiaddr)>

pub fn send_request(&mut self, peer_id: &PeerId, request: T) -> RequestId

pub fn send_response(
&mut self,
response: U,
request_id: RequestId
) -> Result<(), U>

pub fn get_active_mdns_peers(&mut self) -> Vec<&PeerId>

Get the peers discovered by mdns