Module stronghold_communication::behaviour::types

Structs

P2PIdentifyInfo
Information of a peer sent in Identify protocol responses.

RequestEnvelope
Wrapped request message to allow using a relay peer that forward messages from source to target.

Enums

P2PEvent
Event that was emitted by one of the protocols of the P2PNetworkBehaviour

P2PIdentifyEvent
Event emitted by the Identify behaviour.

P2PInboundFailure
Possible failures occurring in the context of receiving an inbound request and sending a response.

P2PMdnsEvent
Event that can be produced by the Mdns behaviour.

P2POutboundFailure
Possible failures occurring in the context of sending an outbound request and receiving the response.

P2PProtocolsHandlerUpgrErr
Error that can happen on an outbound substream opening attempt.

P2PReqResEvent
Event emitted by the RequestResponse behaviour.